[#5] Professor 메뉴 추가
