+++ src/components/views/Professor/Professor.module.css
@@ -1,0 +1,4 @@
+.container {
+  padding: 0px 200px 30px 200px;
+  min-height: calc(100vh);
+}
